requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
viam-sdk
typing-extensions
pyinstaller
